build with opt-lvl 3 and panic unwind
Panic unwind almost doubles the binary size while opt-level 3 when compared to opt `s` increases around 2mb.

The binary size hit with opt-level is not really a problem, specially as opt-level 3 can improve the performance when passing events and messages to the frontend.

The real question is if panic unwind is worth it, I prefer a proper shutdown in any software so I will keep it for a while.